React: Add `<CalendarDateInput>` component
This is like `CalendarInput`, but doesn't specifically accept `Slot`
resources as input. This lets us more easily use it in conjunction with
APIs like `Appointment/$find`.

As a nice side-effect, the existing component probably becomes slightly
more efficient, as the conversion of Slot.start from `string` to `Date`
happens once inside a memoized function. This is not a big deal for 30
days rendered client-side, but still a nice little win.
