+++ packages/react/src/index.ts
@@ -26,6 +26,7 @@ export * from './auth/SetPasswordForm';
 export * from './auth/SignInForm';
 export * from './BackboneElementDisplay/BackboneElementDisplay';
 export * from './BackboneElementInput/BackboneElementInput';
+export * from './CalendarDateInput/CalendarDateInput';
 export * from './CalendarInput/CalendarInput';
 export * from './chat/BaseChat/BaseChat';
 export * from './chat/ChatModal/ChatModal';
